Updates to terragrunt 0.19 syntax
This is still non-working, since this project relies currently on
terragrunt 0.18 functionality, where the hooks executed from the
terragrunt cache directory and so could reference relative paths
to files in the remote modules.

Following commits will update to tf 0.12 versions of the remote
modules, which no longer have that requirement.