Minimizes output to avoid travis-ci limits
Turns out travis-ci has a ~4MB limit on the size of the output log.
When the output exceeds that, the job is cancelled. There is no
workaround, other than reducing the amount of output.

https://docs.travis-ci.com/user/common-build-problems/#log-length-exceeded

This patch suppresses only stdout in the release pipeline, where we've
experienced this problem. Output from the bucket-list is still provided,
as this will generally identify what items will be changing. The exception
is when the yum definition files change. Addressing that would require
larger changes to the pipeline. Prefer to see if this works for us,
first, before making larger changes.

Also, stderr is still output, so we will see the terragrunt output,
as well as any terraform errors if there are problems.